DozeSensors: Only use proximity sensor if supported
On msm-4.14 devices, when the proximity sensor is in use,
the smp2p-sleepstate IRQ is fired multiple times a second,
with each one holding a 200ms wakelock.
This is probably a bug in the DSP firmware.
To fix this, avoid using the proximity sensor in doze mode,
because sleep is preferred to turning off the screen.

[dev-harsh1998]: ForwardPort to R.

Change-Id: Ia55388eff52a2136c87106d81ea295d2fe9140b2
Signed-off-by: Harshit Jain <god@hyper-labs.tech>
Signed-off-by: Akash Srivastava <akashniki@gmail.com>
Signed-off-by: maade <p1nchp0d@bk.ru>